my_flask_app/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
pydantic
email-validator
pytest
apig-wsgi  # Lambda WSGI adapter for API Gateway integration
python-dotenv
psycopg2-binary
PyJWT
supabase
requests
